Corrected type of steeringPWM to prevent division error which ruined PID

+++ raspberryPI3/ros2_ws/src/car_control/src/car_control_node.cpp
@@ -488,7 +488,7 @@ private:
     float smallLeftRearPwmCmd = 0.0;
     float leftRearRPM;
     float rightRearRPM;
-    uint8_t steeringPwmCmd;
+    float steeringPwmCmd;
 
     //gnss data variables
     float currentLatitude;
